+++ modules/AdminToolbox.FortiWizard/Changelog.md
@@ -14,6 +14,7 @@
 * **2.1.1** Added Dependency comments for each function.
 * **2.2.0** Fix for null services parameter on function New-FormTunnel
 * **2.3.0** Fixes for Github issue #26 that address improperly named source and remote address when there is a single source and remore cidr provided.
+* **2.4.0** Updated required modules
 
 #The line below this one is used strictly by a github action. That action is responsible for publishing new versions of this module to Powershell Gallery, upon pushing an iterative version to Github.
-2.3.0
+2.4.0